requirements.txt
Django==5.2.3
djangorestframework==3.16.0
drf-yasg==1.21.10